Util Group 생성
- MyConverter class 생성
- 메소드 옮기기
